acquire admin access if necessary to scan accounts
